Ask for confirmation to apply changes

+++ vehicle-management/vehicle-management.component.js
@@ -7,21 +7,28 @@ angular.module('vehicleManagement').
 
         self.vehicleId = $routeParams.vehicleId;
         $scope.vehicle = null;
-        $scope.displayAlert = false;
+        $scope.displayAlert = {};
+        $scope.displayAlert.statement = false;
+        $scope.displayAlert.name = '';
 
         var getVehicles = JSON.parse(localStorage.getItem('vehicles'));
         $scope.vehicle = getVehicles[self.vehicleId];
 
         $scope.saveChanges = function (vehicle, formValidation) {
-          if (formValidation.$valid) {
+          if (confirm('Are you sure you want to save changes?') && formValidation.$valid) {
             console.log('form is valid.');
             var newData = vehicle;
             var currentData = JSON.parse(localStorage.getItem('vehicles'));
             currentData[self.vehicleId] = newData;
             localStorage.setItem('vehicles', JSON.stringify(currentData));
-            $scope.displayAlert = true;
+            $scope.displayAlert.statement = true;
+            $scope.displayAlert.type = 'alert-primary';
+            $scope.displayAlert.name = 'Data successfully changed!';
           } else {
-            console.log('Form is not valid.');
+            console.log('No Confirmation or Form is not valid.');
+            $scope.displayAlert.statement = true;
+            $scope.displayAlert.type = 'alert-danger';            
+            $scope.displayAlert.name = 'Data not changed!';
           }
         }
       }
